requirements.txt
altair==5.5.0
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.6
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.1.8
einops==0.8.2
filelock==3.19.1
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.46
hf-xet==1.2.0
huggingface_hub==0.36.2
idna==3.11
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
mpmath==1.3.0
narwhals==2.16.0
networkx==3.2.1
numpy==1.26.4
opencv-python==4.13.0.92
packaging==25.0
pandas==2.3.3
pillow==11.3.0
protobuf==6.33.5
pyarrow==21.0.0
pydeck==0.9.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
regex==2026.1.15
requests==2.32.5
rpds-py==0.27.1
safetensors==0.7.0
six==1.17.0
smmap==5.0.2
streamlit==1.50.0
sympy==1.14.0
tenacity==9.1.2
timm==1.0.24
tokenizers==0.22.2
toml==0.10.2
torch==2.2.2
torchvision==0.17.2
tornado==6.5.4
tqdm==4.67.3
transformers==4.57.6
typing_extensions==4.15.0
tzdata==2025.3
urllib3==1.26.20
